add update pets

+++ controllers/pets/addPet.js
@@ -1,46 +1,54 @@
 const { Pet } = require("../../models/pets");
 const { HttpError } = require("../../helpers");
-const Jimp = require("jimp");
+// const Jimp = require("jimp");
 const fs = require("fs/promises");
-// const path = require("path"); // need for avatarDir
-const cloudinary = require("../../helpers/cloudinary");
+// const cloudinary = require("../../helpers/cloudinary");
+const cloudinary = require('cloudinary').v2
 
-const { ObjectId } = require("mongodb");
-
-// const avatarDir = path.join(__dirname, "../../", "public", "pets");
-const imgSizePx = 250;
+// const imgSizePx = 250;
 
 const addPet = async (req, res) => {
   const { _id } = req.user;
-  const newId = new ObjectId(_id);
-
+  if(!req.file) {
+    throw HttpError (404, "Avatar is required")
+  };
   const { path: tempUpload } = req.file;
-  const jimpAvatar = await Jimp.read(tempUpload);
-  await jimpAvatar
-    .resize(imgSizePx, imgSizePx, Jimp.RESIZE_BEZIER)
-    .writeAsync(tempUpload);
+  // const jimpAvatar = await Jimp.read(tempUpload);
+  // await jimpAvatar
+  //   .resize(imgSizePx, imgSizePx, Jimp.RESIZE_BEZIER)
+  //   .writeAsync(tempUpload); // это можно сделать через клаудинари (дополнительная библиотека)
+  // try {
+  //   const uploader = async (path) =>
+  //     await cloudinary.uploads(path);
+  //   const newPath = await uploader(tempUpload);
+  //   fs.unlink(req.file.path);
+  //   const newPet = await Pet.create({
+  //     ...req.body,
+  //     avatar: newPath.url,
+  //     owner: _id, // исправлено
+  //   });
+  //   res.status(201).json({
+  //     message: "success",
+  //     newPet,
+  //   });``
+  // } catch (error) {
+  //   // await fs.unlink(req.file.path); // новый комент
+  //   console.log("====addPet error:", error);
+  //   HttpError(400, error.message);
+  // }
+  const { url } = await cloudinary.uploader.upload(tempUpload);
 
-  try {
-    const uploader = async (path) =>
-      await cloudinary.uploads(path, "petly_dir/pets");
-    const newPath = await uploader(tempUpload);
     fs.unlink(req.file.path);
-
     const newPet = await Pet.create({
       ...req.body,
-      avatar: newPath.url,
-      owner: newId,
+      avatar: url,
+      owner: _id, // исправлено
     });
 
     res.status(201).json({
       message: "success",
-      data: { result: newPet },
+      newPet,
     });
-  } catch (error) {
-    // await fs.unlink(req.file.path);
-    console.log("====addPet error:", error);
-    HttpError(400, error.message);
-  }
 };
 
 module.exports = addPet;
